Add more entries for the statistics simplify the memory statistics remove the temp stats task since this should be done from the application layer

+++ src/main/java/io/mapsmessaging/storage/impl/file/partition/IndexManager.java
@@ -159,7 +159,7 @@ public class IndexManager implements Closeable {
     return (int) counter.sum();
   }
 
-  public int emptySpace(){
+  public long emptySpace(){
     return (int) emptySpace.sum();
   }
